reshuffling before update of date validation
