+++ data/types/time_validation_test.go
@@ -1,0 +1,9 @@
+package types_test
+
+import (
+	. "github.com/onsi/ginkgo"
+)
+
+var _ = Describe("Time Validation", func() {
+
+})
